feat(wip): search modal
